requirements.txt
flask==2.0.1
python-dotenv==0.19.0
flask-sqlalchemy